// src/app/models/models.ts
export interface EventSummary {
  eventId: number;
  name: string;
  type: string | null;
  taxonomyName: string | null;
  taxonomySubName: string | null;
  datetimeUtc: string;
  venueName: string;
  venueCity: string | null;
  venueState: string | null;
  minPrice: number | null;
  ticketsRemaining: number;
  performers: string[];
  imageUrl: string | null;
}